Moves GC logic to update indices
