+++ entity-registry/src/main/java/com/linkedin/metadata/models/registry/ConfigEntityRegistry.java
@@ -122,7 +122,7 @@ public class ConfigEntityRegistry implements EntityRegistry {
       entityNameToSpec.put(entity.getName().toLowerCase(), entitySpec);
     }
 
-    entitySpecBuilder.validateRelationships();
+    //entitySpecBuilder.validateRelationships();
 
     // Build Event Specs
     eventNameToSpec = new HashMap<>();
